+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Not released
 
+### PEXSI
+* AAA method has become the default pole expansion method in PEXSI.
+* Increased default number of poles from 20 to 30.
+* Improved accuracy of pole expansion based on minimax approximation.
+
 ### SLEPc-SIPs
 * Interface compatible with PETSc 3.13 and SLEPc 3.13.